+++ R/modelsum.R
@@ -151,6 +151,11 @@ modelsum <- function(formula,  family="gaussian", data, adjust=NULL, na.action=n
   if(is.null(temp.call$na.action)) {
     temp.call$na.action <- na.modelsum
   }
+  if(!missing(data))
+  {
+    data <- keep.labels(data)
+    temp.call$data <- call("keep.labels", temp.call$data)
+  }
 
   ## if(is.null(temp.call$weights)) {
   ##    temp.call$weights <- rep(1, nrow())
@@ -200,7 +205,7 @@ modelsum <- function(formula,  family="gaussian", data, adjust=NULL, na.action=n
 
   ## create the environment where the formula will be evalulated
   tabenv <- new.env(parent = environment(formula))
-  environment(temp.call$formula) <-  environment(base.call) <- tabenv
+  environment(temp.call$formula) <-  environment(base.call$formula) <- tabenv
 
   basedf <- eval.parent(base.call)
   modeldf <- eval.parent(temp.call)
@@ -497,11 +502,6 @@ modelsum <- function(formula,  family="gaussian", data, adjust=NULL, na.action=n
     fitList[[xname]]$glance <- c(fitList[[xname]]$glance, endpoint=yTerm, endlabel=yLabel, x=xname)
 
   } # end for: eff
-
-  ##  if(!usingRCF() & !usingNCSA()) {
-  ##    cat(paste0("R-", version$major,".", version$minor, "\t", system("echo $USER",intern=TRUE), "\t", Sys.Date(), "\n"),
-  ##        file="/projects/bsi/infrastructure/s200555.Rinfrastructure/rlogs/modelsum.log",append=TRUE)
-  ##  }
 
   msList <- list(fits=fitList, control = control, Call = Call, family=family)
                  ##, adjust=adjVars)