+++ R/tableby.stat.tests.R
@@ -33,7 +33,7 @@ chisq <- function(x, x.by) {
   tab <- table(x, x.by, exclude=NA)
   ctl <- dynGet("control") # envir=parent.frame)
   if(sum(rowSums(tab)>0)>1) {
-    stats::chisq.test(tab[rowSums(tab)>0,], correct=ctl$chisq.correct, simulate.p.value=ctl$simulate.p.value, B=ctl$B)
+    suppressWarnings(stats::chisq.test(tab[rowSums(tab)>0,], correct=ctl$chisq.correct, simulate.p.value=ctl$simulate.p.value, B=ctl$B))
   } else {
     list(statistic=0, p.value=1, method="Pearson's Chi-squared test")
   }